+++ libraries/cmsis/dsp/Source/StatisticsFunctions/arm_logsumexp_f32.c
@@ -3,11 +3,13 @@
  * Title:        arm_logsumexp_f32.c
  * Description:  LogSumExp
  *
+ * $Date:        23 April 2021
+ * $Revision:    V1.9.0
  *
  * Target Processor: Cortex-M and Cortex-A cores
  * -------------------------------------------------------------------- */
 /*
- * Copyright (C) 2010-2019 ARM Limited or its affiliates. All rights reserved.
+ * Copyright (C) 2010-2021 ARM Limited or its affiliates. All rights reserved.
  *
  * SPDX-License-Identifier: Apache-2.0
  *
@@ -24,13 +26,13 @@
  * limitations under the License.
  */
 
-#include "arm_math.h"
+#include "dsp/statistics_functions.h"
 #include <limits.h>
 #include <math.h>
 
 
 /**
- * @addtogroup groupStats
+ * @addtogroup LogSumExp
  * @{
  */
 
@@ -108,10 +110,10 @@ float32_t arm_logsumexp_f32(const float32_t *in, uint32_t blockSize)
        tmp = *pIn++;
        accum += expf(tmp - maxVal);
        blkCnt--;
-
+    
     }
 
-    accum = maxVal + log(accum);
+    accum = maxVal + logf(accum);
 
     return (accum);
 }
@@ -133,7 +135,7 @@ float32_t arm_logsumexp_f32(const float32_t *in, uint32_t blockSize)
     const float32_t *pIn;
     uint32_t blkCnt;
     float32_t accum;
-
+ 
     pIn = in;
 
     blkCnt = blockSize;
@@ -146,7 +148,7 @@ float32_t arm_logsumexp_f32(const float32_t *in, uint32_t blockSize)
       while(blkCnt > 0)
       {
          tmp = *pIn++;
-
+  
          if (tmp > maxVal)
          {
             maxVal = tmp;
@@ -164,7 +166,7 @@ float32_t arm_logsumexp_f32(const float32_t *in, uint32_t blockSize)
       {
          tmpVb = vld1q_f32(pIn);
          pIn += 4;
-
+  
          idxV = vcgtq_f32(tmpVb, maxValV);
          maxValV = vbslq_f32(idxV, tmpVb, maxValV );
 
@@ -180,7 +182,7 @@ float32_t arm_logsumexp_f32(const float32_t *in, uint32_t blockSize)
       while(blkCnt > 0)
       {
          tmp = *pIn++;
-
+  
          if (tmp > maxVal)
          {
             maxVal = tmp;
@@ -190,7 +192,7 @@ float32_t arm_logsumexp_f32(const float32_t *in, uint32_t blockSize)
 
     }
 
-
+    
 
     maxValV = vdupq_n_f32(maxVal);
     pIn = in;
@@ -208,7 +210,7 @@ float32_t arm_logsumexp_f32(const float32_t *in, uint32_t blockSize)
        accumV = vaddq_f32(accumV, tmpV);
 
        blkCnt--;
-
+    
     }
     accumV2 = vpadd_f32(vget_low_f32(accumV),vget_high_f32(accumV));
     accum = vget_lane_f32(accumV2, 0) + vget_lane_f32(accumV2, 1);
@@ -219,7 +221,7 @@ float32_t arm_logsumexp_f32(const float32_t *in, uint32_t blockSize)
        tmp = *pIn++;
        accum += expf(tmp - maxVal);
        blkCnt--;
-
+    
     }
 
     accum = maxVal + logf(accum);
@@ -234,7 +236,7 @@ float32_t arm_logsumexp_f32(const float32_t *in, uint32_t blockSize)
     const float32_t *pIn;
     uint32_t blkCnt;
     float32_t accum;
-
+ 
     pIn = in;
     blkCnt = blockSize;
 
@@ -250,7 +252,7 @@ float32_t arm_logsumexp_f32(const float32_t *in, uint32_t blockSize)
           maxVal = tmp;
        }
        blkCnt--;
-
+    
     }
 
     blkCnt = blockSize;
@@ -261,7 +263,7 @@ float32_t arm_logsumexp_f32(const float32_t *in, uint32_t blockSize)
        tmp = *pIn++;
        accum += expf(tmp - maxVal);
        blkCnt--;
-
+    
     }
     accum = maxVal + logf(accum);
 
@@ -271,5 +273,5 @@ float32_t arm_logsumexp_f32(const float32_t *in, uint32_t blockSize)
 #endif /* defined(ARM_MATH_MVEF) && !defined(ARM_MATH_AUTOVECTORIZE) */
 
 /**
- * @} end of groupStats group
+ * @} end of LogSumExp group
  */
